Feature/lolog (#49)
* fix: 소개글 조회/작성

소개글 조회
- 소개글을 작성한 사용자와 로그인한 유저가 일치 여부 추가

소개글 작성
- 소개글 작성 시 로그인한 사용자의 token 값으로 작성되도록 수정

* refactor: 롤로그 소개 조회/편집
lologController에 있던 소개 조회/편집 기능을 aboutController로 따로
분리하였음.

* refactor: 롤로그 시리즈 조회
롤로그 시리즈 조회 시에 seriesRepository를 통해서 가지고 오도록 변경.
그 과정에서 다른 moduel들의 불필요한 exports 제거.
+) series 서비스 및 레포지토리의 오타 수정

* fix: 시리즈 조회
시리즈 조회 시 create_at 이 아닌 update_at을 읽어오도록 수정

* fix: 소개글 API 주소
소개글 API 주소 수정

* refactor: 시리즈 조회
시리즈 조회 기능을 seriesController로 옮겼음.
그 과정에서 시리즈의 디테일을 조회하는 api주소가 변경되었음.

* fix: 롤로그 불러오기
게시글과 태그 목록을 불러올 때 해당 롤로그의 소유자 정보도 같이
불러오도록 수정

+++ src/repository/user.repository.ts
@@ -33,8 +33,7 @@ export class UserRepository extends Repository<User> {
   }
 
   async signupWithSocial(createSocialUserDto: CreateSocialUserDto) {
-    const { name, email, login_id, about_me, provider, profile_image } =
-      createSocialUserDto;
+    const { name, email, login_id, about_me, provider, profile_image } = createSocialUserDto;
     const user = this.create({
       name,
       email,
@@ -49,18 +48,11 @@ export class UserRepository extends Repository<User> {
   }
 
   async updateUser(id: number, updateData: object) {
-    await this.createQueryBuilder()
-      .update()
-      .set(updateData)
-      .where('id = :id', { id })
-      .execute();
+    await this.createQueryBuilder().update().set(updateData).where('id = :id', { id }).execute();
   }
 
   async getUserByUserId(id: number, keys: string[]) {
-    return await this.createQueryBuilder()
-      .select(keys)
-      .where('id = :id', { id })
-      .execute();
+    return await this.createQueryBuilder().select(keys).where('id = :id', { id }).execute();
   }
 
   async updateProfileImage(id: number, profile_image: string) {
@@ -100,17 +92,14 @@ export class UserRepository extends Repository<User> {
       .execute();
   }
 
-  async selectAboutBlog(user_id: number) {
+  async selectAboutBlog(user_id: number, login_user_id: number) {
     return await this.query(
-      `SELECT id AS user_id, about_blog FROM user WHERE id = ?`,
-      [user_id],
+      `SELECT id AS user_id, about_blog, IF(id = ?, 1, 0) as is_owner FROM user WHERE id = ?`,
+      [login_user_id, user_id],
     );
   }
 
   async withdrawal(user_id: number) {
-    await this.createQueryBuilder()
-      .delete()
-      .where('id = :user_id', { user_id })
-      .execute();
+    await this.createQueryBuilder().delete().where('id = :user_id', { user_id }).execute();
   }
 }
